requirements.txt
# Core ML + Embeddings
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.39.3

# Vector Database
faiss-cpu==1.8.0

# API + Web
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2

# Utilities
pandas==2.2.1
tqdm==4.66.2
python-multipart==0.0.9
pydantic==2.7.0
scikit-learn==1.4.1.post1

# OpenAI (optional for LLM generation)
openai==1.14.3
